pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi (==0.115.8)",
    "uvicorn (==0.34.0)",
    "py-eureka-client (==0.11.13)",
    "python-dotenv (==1.0.1)",
    "langchain-core (==0.3.72)",
    "langgraph (==0.2.70)",
    "langchain-mistralai (==0.2.6)",
    "langchain-openai (==0.3.11)",
    "langchainhub (==0.1.21)",
    "langchain-community (==0.3.17)",
    "langchain (==0.3.18)",
    "psycopg2-binary (==2.9.10)",
    "langchain-google-genai (==2.1.4)",
    "jinja2 (==3.1.6)",
    "langgraph-checkpoint-redis (==0.0.8)",
    "beautifulsoup4 (==4.13.4)",
    "sentence-transformers (==5.0.0)",
    "langchain-huggingface (==0.3.1)",
    "langchain-text-splitters (==0.3.9)",
    "langchain-pinecone (==0.2.11)",
    "pinecone (>=4.0.0)",
    "scrapy-playwright (==0.0.43)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
